refactor files by ops in test folder
